Fix interaction between provisional nodes and mark.persist.
